requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
twilio==8.10.0
openai==1.54.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
pytz==2023.3
httpx==0.27.0
pytest-asyncio==0.21.1
pytest==8.2.2
